local-requirements.txt
apache-airflow==2.9.2
apache-airflow-providers-postgres
apache-airflow-providers-http
pandas
